requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==12.0
pydantic==2.8.2
numpy==1.26.4
xxhash==3.4.1
google-genai==0.3.0
